// src/api/generated/models/postCompaniesCompanyIdImportsRequestLinesHtmPreview200PreviewRowsItem.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { PostCompaniesCompanyIdImportsRequestLinesHtmPreview200PreviewRowsItemData } from './postCompaniesCompanyIdImportsRequestLinesHtmPreview200PreviewRowsItemData';
import type { PostCompaniesCompanyIdImportsRequestLinesHtmPreview200PreviewRowsItemParsed } from './postCompaniesCompanyIdImportsRequestLinesHtmPreview200PreviewRowsItemParsed';

export type PostCompaniesCompanyIdImportsRequestLinesHtmPreview200PreviewRowsItem = {
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  rowNumber: number;
  data: PostCompaniesCompanyIdImportsRequestLinesHtmPreview200PreviewRowsItemData;
  errors: string[];
  parsed?: PostCompaniesCompanyIdImportsRequestLinesHtmPreview200PreviewRowsItemParsed;
};