// src/api/generated/models/postCompaniesCompanyIdImportsRequestLinesTranslatePreview200Preview.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { PostCompaniesCompanyIdImportsRequestLinesTranslatePreview200PreviewRowsItem } from './postCompaniesCompanyIdImportsRequestLinesTranslatePreview200PreviewRowsItem';

export type PostCompaniesCompanyIdImportsRequestLinesTranslatePreview200Preview = {
  rows: PostCompaniesCompanyIdImportsRequestLinesTranslatePreview200PreviewRowsItem[];
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  validRowCount: number;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  invalidRowCount: number;
};