// src/api/generated/models/postCompaniesCompanyIdImportsRequestLinesTranslatePreviewBodyPreview.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { PostCompaniesCompanyIdImportsRequestLinesTranslatePreviewBodyPreviewRowsItem } from './postCompaniesCompanyIdImportsRequestLinesTranslatePreviewBodyPreviewRowsItem';

export type PostCompaniesCompanyIdImportsRequestLinesTranslatePreviewBodyPreview = {
  rows: PostCompaniesCompanyIdImportsRequestLinesTranslatePreviewBodyPreviewRowsItem[];
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  validRowCount: number;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  invalidRowCount: number;
};