// src/api/generated/models/postCompaniesCompanyIdRequestsBodyLinesItem.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { PostCompaniesCompanyIdRequestsBodyLinesItemAttributes } from './postCompaniesCompanyIdRequestsBodyLinesItemAttributes';

export type PostCompaniesCompanyIdRequestsBodyLinesItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  productId?: string;
  /** @minLength 1 */
  description: string;
  /** @pattern ^\d+(\.\d{1,4})?$ */
  quantity: string;
  /** @minLength 1 */
  unit?: string;
  /** @minLength 1 */
  sku?: string;
  /** @minLength 1 */
  notes?: string;
  attributes?: PostCompaniesCompanyIdRequestsBodyLinesItemAttributes;
};